// library/src/_baseGame/types/ConnectionStatus.ts
/*
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

export enum ConnectionStatus {
  Unknown = 0,
  Disconnected = 1,
  Connecting = 2,
  Connected = 3,
  Disconnecting = 4,
  Offline = 5
}